:bug: Update child occupancy check to use decimal type.
Children apparently can be partials as they count a child who is present part time as 0.5 a child.

+++ api/src/models/funding-submission-line-json.ts
@@ -143,7 +143,7 @@ export class FundingSubmissionLineJson extends BaseModel<
                   JSON_VALUE(
                     json_array_element.value,
                     '$.actualChildOccupancyRate'
-                  ) AS int
+                  ) AS decimal(10,2)
                 ),
                 0
               )
